Refactor slider
