+++ src/components/GalleryPage/Gallery.js
@@ -1,20 +1,8 @@
 import React from 'react';
 import { graphql, useStaticQuery } from 'gatsby';
-import Img from 'gatsby-image';
-import useTheme from '@material-ui/core/styles/useTheme';
-import useMediaQuery from '@material-ui/core/useMediaQuery';
-import Slider from '../CarouselSlider/CarouselSlider';
+import CarouselSlider from '../CarouselSlider/CarouselSlider';
 
 export default function Gallery() {
-	const settings = {
-		dots: true,
-		infinite: true,
-		speed: 800,
-		slidesToShow: 1,
-		slidesToScroll: 1,
-		className: 'slides'
-	};
-
 	const data = useStaticQuery(graphql`
 		{
 			carouselImages: allFile(filter: {relativePath: {glob:"gallery-carousel/*"}}) {
@@ -31,17 +19,7 @@ export default function Gallery() {
 		}
 	`);
 
-	const theme = useTheme();
-	const wideImg = useMediaQuery(theme.breakpoints.up('md'));
-	const imageWidth = wideImg ? 2.5 : 2;
-
-	const images = data.carouselImages.nodes.map(node => {
-		return <Img fluid={{ ...node.childImageSharp.fluid, aspectRatio: imageWidth }} key={node.id} />;
-	});
-
 	return (
-		<Slider {...settings}>
-			{images}
-		</Slider>
+		<CarouselSlider imageNodes={data.carouselImages.nodes} />
 	);
 }
